Add PgBouncer peering support (#666)
For both performance and availability reasons it can be useful to load
balance across multiple PgBouncer processes, possibly on different
machines. It's very easy to do, by simply putting a generic TCP load
balancer in front of these pgbouncer processes. The `so_reuseport`
option is an example of such a load balancer, which uses the OS kernel
to do the load balancing.

However, there is a problem with this simple approach. Query
cancellations will not work most of the time. The reason is that a
cancellation will open another TCP connection, which might be handled by
another pgbouncer process than the one that is processing the query that
should be cancelled. I talked about this problem in more detail at
the Uptime conference: https://www.youtube.com/watch?v=M585FfbboNA

This change adds support for pgbouncer peering. To enable peering two
things need to be added to the configuration of each pgbouncer:
1. A peer_id that's unique for each pgbouncer that is peered with
2. A new "peers" section that should get an entry for each of the other
   peers

When pgbouncer is configured like this, then whenever it receives a
cancellation request for an unknown session, it will forward it to the
peer that owns this session.

Co-authored-by: Marco Slot <marco.slot@gmail.com>