Fail hard when UNIX socket cannot be created (#830)
When a user requests that a UNIX socket is created and this fails for
some reason we ignored the error. This starts failing hard when that
happens.

This problem is especially bad when so_reuseport is used in combination
with peering, because the peer list would then contain sockets that
don't exist and forwarding cancels would wait for timeout and then fail
silently.