Fail hard if a listen_addr could not be listened on (#838)
This fails if listen_addr was specified, but none of the addresses could
be listened on successfully. While not being able to listen on some of
the addresses is often acceptable in practice, not being able to listen
on any of them clearly indicates an error.

This behavior is the same as Postgres its behavior.