Fix FATAL in function server_proto(): server in bad state: 14
In most cases the postgres server will not respond anything back to a
cancel message. It simply closes the connection. It won't even report an
error if the cancellation key is incorrect, or the PID is unknown.
However, it is possible that it sends another type of error message,
such as an OOM. Libpq actually ignores those errors and won't show them
to the user, but they might still be sent by the server. Because of the
changes in #717 this could cause a FATAL error. This fixes that by now
handling `SV_ACTIVE_CANCEL` the same as `SV_ACTIVE` in `server_proto`.

Fixes #847

Related to #717, #801, and #815