Retry adding an editorconfig file (#841)
This allows many editors to be configured with correct indentation
styles for the different filetypes that we have. Github also parses this
file and uses it for its web-based editor as well as for displaying
files.

It's a slightly modified version of reverted commit a718ad8. This new
version doesn't specify tab_width and indent_size for C files.