Use unresolvable DNS rather than bad IP to test host_strategy
Previously, the host_strategy tests used 127.0.0.2,127.0.0.1 had a "bad
IP" (127.0.0.2) as the first entry to force the first backend server
connection to fail and verify that `host_strategy=last_successful`
directs new connections to the second "good" entry.

MacOS doesn't treat 127.0.0.2 in the same manner as Linux by default, so
we need a different way to fail the first host in the list. Switching to
an unresolvable DNS hostname works fine on both Linux and MacOS.