Provide SQLSTATE (#814)
PgBouncer currently uses SQLSTATE 08P01 (protocol_violation) as default
error code to send error messages to client. There are cases that this
SQLSTATE is different from what Postgres provides. Since the goal is to
be as transparent as possible, PgBouncer should report the SQLSTATE
provided by Postgres. Issue #778 exposes one case that if you connect to
a database that doesn't exist, Postgres reports 3D000
(invalid_catalog_name) but PgBouncer reports 08P01. The npgsql driver
relies on the SQLSTATE to check if the database already exists and this
behavior breaks it. PgBouncer will use the SQLSTATE that Postgres
provides.

Fixes #778