Add support for options startup parameter (#878)
It's allowed to specify any Postgres command line parameter in the `options`
startup packet #867. This is commonly used to specify GUC values on connection
startup, using the `-c key=value` command line flag. I believe the main reason
this is used instead of specifying the desired GUC as a startup parameter
directly is because libpq does not allow specifying arbitrary startup
parameters.

This PR adds support for configuring GUCs using the `-c` flag in the options
startup packet. While in theory people can specify any flag they want. In
practice the `-c` flag is the only one that I know is commonly used. Probably
because it is the only one that's documented in the Postgres docs for
`PGOPTIONS`. Any of the other command line flags are simply shorthands for
specific GUCs, and so even when those would be used they can be easily rewritten
using `-c`. So to keep parsing of the options startup parameter simple, this
only supports the `-c` flag.

Fixes #875