refactor: Use an enum for shutdown values
In #902 I want to introduce a 4th shutdown value. So I started using an
enum for this to improve readibility. I think this refactor from using
integer values to enum values stands on its own. So to make reviewing
easier, and not have it be blocked on adding tests and some discussions
for #902 this PR includes only that refactor.