Fix confusing typo TLS/TCP typo in docs
This was originally reported by @ntse on the repo for the website. And
it was merged in https://github.com/pgbouncer/pgbouncer.github.io/pull/9

During the last release process I noticed that the change was
automatically overwritten, because for `config.md` this repo is the
canonical place. So this does the change again on the right repo.

Co-authored-by: Sam Ainsworth <samuel.ainsworth@digital.justice.gov.uk>