Add fallback option to auth_dbname
When the database section has an auto-database, '*', the expected
behavior for login experience to have client's database registered
when not found. However, this experience would be blocked if we
configure auth_dbname.

This PR makes already existing function, find_or_register_database,
public and uses it from various places to simplify code.