Add autoformatting for C files (#892)
This uses uncrustify to autoformat C files. It adds a config that matches
the code style of the repo as close as I was able to configure it.

We also considered using pgindent, but that requires a typedefs.list which is a
pain to create. The current codestyle of PgBouncer also doesn't match the one
from Postgres at all, so it would cause many uninteresting changes.