Use server_tls_sslmode=disable in some flaky tests
In 1.20.0 we started using `server_tls_sslmode=prefer` by default. This
sets it to disable for specific tests that depend heavily on specific
timings. These timings would be messed up on some OSes in CI, because
now PgBouncer would first try to connect using TLS and when that failed
it would try connect without. Especially on Windows CI a failed TLS
connection to localhost seems to sometimes take significant time.