Use auto-database when database in auth_dbname is not explicitly configured (#921)
When the database section has an auto-database, '*', the expected
behavior for login experience to have client's database registered
when not found. However, the auto-database wasn't used when
the database specified in auth_dbname wasn't explicitly configured.
This patch introduces the behaviour that we also fallback to the
auto-database in that case.