Improve security of OpenSSL settings (#948)
The default OpenSSL ciphers were extremely outdated. This PR brings them
up to date. The main code changes for this are in libusual (which this
PR updates):
https://github.com/libusual/libusual/pull/40

The changes in PgBouncer itself are mostly doc changes. The only code
change is changing the default of the cipher to actually be "default",
which now means the same as "fast" anyway.