PgBouncer md5 FIPS compatibility changes (#949)
MD5 encryption functions are internally implemented in libusual.
MD5 Authentication functions are not allowed to be used on FIPS
compliant host. But due to this implementation of MD5 algorithm
MD5 authentication it is working fine on FIPS enabled host. Instead
it should throw an error.

This is fixed by replacing the calls to libusual its MD5 implementation
with calls to OpenSSL MD5 APIs (when compiled with OpenSSL support).
By doing this a MD5 authentication now fails when running with a FIPS
compliant OpenSSL version.