Don't kill all waiting clients on temporary FATAL errors during login (#946)
If the server connection returns a FATAL error in the login phase we
kill all the waiting clients and show this error to the user. The reason
is that when this happens it's usually because of some configuration
error (e.g. pg_hba.conf on Postgres not being configured correctly).

But there are common cases where the error is temporary too, most
importantly: "the database system is starting up"

Postgres returns a specific error code for these types of errors. This
PR starts checking for that error code, and won't disconnect the clients
when it received a FATAL error from the server with that specific error
code.

The primary reason I spent time to fix this is because it was making
our tests flaky:
https://cirrus-ci.com/task/5607330616705024