Maintain `min_pool_size` for pools with forced user (even if no client conns) (#947)
PGBouncer only creates a pool for a user-database combination if at least 1
client is connected as that user to that database.

This means that `min_pool_size` will only be respected if clients are
connected. If all clients disconnect, then the pool can drop below the
`min_pool_size` setting to `0` (`server_lifetime` will eventually close all
connections). This can cause performance problems when a client finally
reconnects since all of the pool connections have to be recreated.

This PR changes that behavior for databases that have forced users. For
databases with forced users pools will now be created automatically even when
there are no connections. This means that for these databases `min_pool_size`
will be respected even without any clients connected.

Partially fixes #426

Continuation of #437