Improve documentation
Fix a few typos. Rewrite some sentences. Replace the future tense with
the present tense in a few sentences.

Remove a reference to an in-progress patch for PostgreSQL that will
avoid an issue with this feature. We should add a reference to it when
such feature is merged in Postgres.