Add timestamp to SHOW commands
Postgres has a facility to include the function now() while collecting
statistics from it. Unfortunately PgBouncer does not have this option.
When we are collecting statistics from PgBouncer, the collector
generally obtain the current timestamp and attach this information to
the statistics that were collected. Instead of relying on this
additional step, this change includes a new column (now) that
already provides the timestamp. The following commands were modified:

- SHOW DATABASES
- SHOW POOLS
- SHOW CLIENTS
- SHOW SERVERS
- SHOW USERS
- SHOW SOCKETS
- SHOW ACTIVE_SOCKETS
- SHOW STATS
- SHOW STATS_TOTALS
- SHOW STATS_AVERAGES