Fix false positive reported by valgrind when using hba (#1001)
To make `valgrind` pass we have a cleanup function that we only use in
assert mode. In this function we free all globals before process exit.
However, we didn't free `parsed_hba`. This fixes that. In passing it
also slightly modifies how we free `parsed_hba` when reloading the
config.