Fix COPY FROM STDIN handling (#1025)
In #845 full tracking of expected responses was implemented. This turned
out to have bugs when sending `COPY FROM STDIN` queries using the
extended protocol. [The protocol explicitly allows sending `Sync`
messages after sending an `Execute` message for a `COPY FROM STDIN`
command][1]. Such Sync messages are ignored by the server until a
CopyDone or CopyFail message is received. The new command queue tracking
did not take this into account, and was incorrectly expecting more
ReadyForQuery messages from the server.

This could lead to two problems for the server connection on which the
`COPY FROM STDIN` was sent:
1. The server connection would never be unassigned from the client.
2. Memory usage for this server connection would continue to increase
until the client disconnects.
3. If `max_prepared_statements` was non-zero, then clients might receive
ParseComplete and CloseComplete responses in the wrong order. This
wouldn't result in wrong behaviour by the server, but the client might
get confused and close the connection.

This PR fixes these issues. It also removes the rfq_count tracking,
since it's now only keeping track of a strict subset of what the
outstanding request queue was tracking.

[1]: https://www.postgresql.org/docs/current/protocol-flow.html#PROTOCOL-COPY

Fixes #1023
Fixes #1024