Update FreeBSD and OSX images for CI (#1030)
Builds have started failing since the previous images don't exist
anymore. Also disables SSL tests on OSX for now, because these
seem to be broken for some reason on the new images. Fixing
those tests is considered out of scope for this PR, but it is
tracked in #1031 (because it should be fixed).