Add values in ignore_startup_parameters
Currently you can only define the key of a parameter in the
ignore_startup_parameters section. With this change you can define a
key=value string in that section. For example you can set the following:

[pgbouncer]
ignore_startup_parameters = options,extra_float_digits=2

If an application attempts to connect using as startup parameter the
extra_float_digits=3, then pgBouncer will produce an error.

This is useful to avoid hidden bugs when using this section between
application logic changes. Also it can function for clients as
indicatory of the expected startup parameters.

Signed-off-by: Marjus Cako <marios.cako@hpe.com>