Improves README (#1051)
The README uses ... in the configure command line to indicate that
options can be informed. However, it can confuse people (like in the
referred issue #1050). Removes the 3 dots and explains that configure
options can be informed.

Since I'm at it, informs that uthash submodule is also fetched if you
are following the Git instructions.

Fixes #1050