Correctly handle partial startup packets
We were bailing out from a connection if we did not receive the full
startup packet in one go, or if it didn't fit in pkt_buf. This fixes
these issues by reusing the packet buffering logic that was introduced
for prepared statement support for our startup packet parsing.

Fixes #1056