Add `server_lifetime` to databases config (#1057)
This PR makes it possible to specify `server_lifetime` on a per database
basis in the `[databases]` section.

This can be useful when a pgbouncer instance that pools connections
for several different severs, of which some makes heavy use of partitioning.
The partitioning causes connections to become very memory heavy
overtime and by lowering server_lifetime these connections can be refreshed
more often. Without this change that could only be done by lowering
server_lifetime globally and thus incurring the overhead of re-authenticating
even for servers where no partitioning is used.