Redesign shutdown logic to support rolling restarts using so_reuseport (#902)
In 1.20 we announced deprecation of online restarts using the `-R` flag
and recommended for people to use `so_reuseport` instead. But the safe
shutdown logic that we executed when receiving a `SIGINT` signal didn't
actually allow for zero-downtime rolling restarts.

This PR changes the behaviour of our shutdown signals to allow for
rolling restarts. The changes that this PR makes are:

1. SIGTERM does not do "immediate shutdown" anymore, but instead
triggers a newly introduced "super safe shutdown". This "super safe
shutdown" waits for all clients to disconnect before shutting down the
server. This can be used for rolling restarts as described in the newly
added documentation. A second SIGTERM will cause an "immediate
shutdown".
2. SIGINT mostly keeps doing what it was doing previously: waiting until
all servers are released before shutting down PgBouncer. But it
incorporates a few improvements from the new SIGTERM logic:
1. It also stops listening for new connections while PgBouncer is
waiting for servers to be released. This does have the downside that
SIGUSR2 cannot be used to cancel a safe shutdown anymore, but that was
quite a weird/niche feature anyway.
2. Clients that have released their server and then do another query
will be disconnected, instead of putting them in pause mode until either
`query_wait_timeout` triggers or PgBouncer shuts down. This way they can
reconnect quicker to another PgBouncer, because they were never going to
get a new server anyway.
     3.  A that a second SIGINT will trigger an "immediate shutdown"
3. SIGQUIT handling is introduced and now causes an immediate shutdown.

Since this changes the existing behaviour of SIGTERM and SIGINT, this
can be considered a breaking change. But the amount of breakage should
be minimal since most of the time people would not want an immediate
shutdown, and the new SIGINT behaviour is closer to the behaviour most
people would expect.

Related to #894
Related to #900

Since this changes SIGTERM to not do a fast shutdown anymore this also
indirectly addresses #806 and #770.

Fixes #806 
Fixes #770