Support --config=value format in options too
The `-c config=value` pattern has been supported in PgBouncer since
PR #878. As described in #952 Postgres also allows to use the
`--config=value`, which honestly seems more readable, but for some
reason is mentioned almost nowhere in the Postgres documentation.

Since #878 this always throws an error, and putting `options` in
`ignore_startup_parameters` wouldn't get rid of this error. So this
should actually be considered a regression for the same reason as #907
was considered a regression (although a much less impactful one given
we've only received two reports of this causing issues).

Fixes #952