ci: Trigger some tasks manually
Reactive some tasks that were commented out in 92b4645.  Instead, set
them to trigger manually.  That way, they can still be used
individually when desired.

Unfortunately, this does not appear to work for matrix elements inside
a task, only for the whole task.  So this only reactives the "Linux
(Red Hat)" and "Linux (Alpine)" tasks, but does not reactivate the
commented out Ubuntu/Debian and Windows variants.