get_pool works with global pool_list (#1103)
The reason for #1103 is that pgbouncer can have multiple credentials with one name.

So we cannot to compare credentials by pointers, we have to check they names.

The second, credentials must not have a local list of pools. get_pool must to work with global list of pools and must check credentials names, too.

This commit corrects get_pool and cmp_pool to apply a new behaviour.