Exact compare function for PgGlobalUser::credentials::tree_node
user_tree stores PgGlobalUser objects but uses credentials_node_cmp that oriented on PgCredentials.

Of course, it works now because PgGlobalUser::credentials is a first member of PgGlobalUser.

If we do not want to trap to hell for this hard coded structure of PgGlobalUser, let's define and use a correct compare function - global_user_node_cmp.

Amen.