Fix for #1104 [free PgCredentials] (#1105)
After the refactoring of the user managament in #1052, we incorrectly
freed credentials by calling `slab_free` on a user_cache, instead of
`credentials_cache` in two places. This could later cause segfaults
because ththe user type is larger than the credentials type, so out of
bound reads would be possible when the credential that was put into the
user_cache was taken out again and used as a user instead.