Make PostgreSQL listen on both IPv4 and IPv6
On certain test environments, there can be unconfigurable hosts where
IPv6 takes precedence over IPv4. This leads PgBouncer to PostgreSQL
connections to be opened in IPv6 traffic, where PostgreSQL doesn't
accept such communication since it is not configured to do so.

This commit adds a line in postgresql.conf file during tests to have it
listen both if the environment variable is set.