Make PostgreSQL listen on both IPv4 and IPv6 during tests if IPv6 interface is supported (#1122)
On certain test environments, there can be unconfigurable hosts where
IPv6 takes precedence over IPv4 (normally, one could control that
through /etc/gai.conf, but if the environment is restricted, it is not
possible.) This leads PgBouncer to PostgreSQL connections to be opened
in IPv6 traffic, where PostgreSQL doesn't accept such communication
since it is not configured to do so. Eventually, the test times out. 

This change adds a line in postgresql.conf file during test startup to have
it listen both IPv4 and IPv6 if the environment variable is set.