Exact compare function for PgGlobalUser::credentials::tree_node (#1109)
user_tree stores PgGlobalUser objects but used credentials_node_cmp that
oriented on PgCredentials. Of course, it works because 
PgGlobalUser::credentials is the first member of PgGlobalUser.

But it can look a bit confusing and in the future maybe we'd like another
field to be first. So to avoid any problems in the future this creates a
dedicated compare function for PgGlobalUser.